third_party/blink/web_tests/virtual/popover-hint-new-behavior-disabled/external/wpt/html/semantics/popovers/popover-top-layer-nesting-hints-expected.txt

This is a testharness.js-based test.
[FAIL] Nested auto/hint ancestors, target is auto with dialog
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Nested auto/hint ancestors, target is auto with fullscreen
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Unrelated hint, target=auto with dialog
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Unrelated hint, target=auto with fullscreen
  assert_equals: Incorrect behavior expected true but got false
Harness: the test ran to completion.
